Start of decoding python scripts.
As is the case with NetEase's other Game Engine (NeoX) most of
the games code is written in python.
All files on disk that contain E2 06 as the first 2 bytes appear to be
those python scripts.
In this particular case, slightly mangled and zlib compressed.

This is the first step in decoding them, which handles the
de-mangeling and zlib decompress.
Next steps is mapping out all the custom and shuffled opcodes.
(This will be quite time intensive, as it looks like there are quite a
few custom opcodes in play here).

So the current state is, we have python-ish codeobjects dumped to disk!